Consolidated iconifying functions to classes
